Update index.md
Hello, I'm Dounia Latorre, a French student at a Master's degree in Translation listed by the European Commission (EMT label). Translation is part of my daily life. I came across your Guide while surfing in the Internet. I saw that it was editable on GitHub. That's why I would like to suggest some modifications. The text would sound more "French". I leave you the choice to accept them or not. In case you are interested, I can commit myself to reread all the pages and modify them if necessary. 
The subject interests me a lot and I feel concerned indirectly. A member of my family is blind. Over time, I realized that we can't know what that means without living with a person with a disability. There are a lot of differences and going out with a person with a disability has some responsibility. My uncle has a dog, adapted electronic devices (phone, computer, GPS...) and I think there is still a lot to learn about the subject. 
I am very motivated to contribute to your project. I may be able to translate it into Spanish later on if you are interested. 

Hoping to receive an answer soon. 

Miss Latorre Dounia